feat: separate out models
