+++ ScatterSim/LatticeObjects.py
@@ -1,3 +1,4 @@
+# Original author : Kevin G. Yager (most of these lattices are Kevin G. Yager)
 from copy import deepcopy
 import numpy as np
 
@@ -467,600 +468,3 @@ class Lattice:
             s += "           \t%s (%s)\n" % (objstr, obj.to_short_string())
             s += "           \t  at pos = (%.3f,%.3f,%.3f)\n" % (xi, yi, zi)
         return s
-
-# Different Lattices
-
-# TODO : The next three should inherit a SimpleCubic lattice
-
-
-class SimpleCubic(Lattice):
-    def __init__(
-            self,
-            objects,
-            lattice_spacing_a=1.0,
-            sigma_D=0.01,
-            lattice_coordinates=None,
-            lattice_types=None):
-        # prepare variables of lattice
-        symmetry = {
-            'crystal family': 'cubic',
-            'crystal system': 'cubic',
-            'Bravais lattice': 'P',
-            'crystal class': 'hexoctahedral',
-            'point group': 'm3m',
-            'space group': 'Pm3m',
-        }
-        if not isinstance(objects, list):
-            objects = [objects]
-
-        lattice_positions = ['placement'] * len(objects)
-
-        if lattice_coordinates is None:
-            lattice_coordinates = [(0.0, 0.0, 0.0)] * len(objects)
-        lattice_types = list(range(len(objects)))
-
-        # now call parent function to initialize
-        super(
-            SimpleCubic,
-            self).__init__(
-            objects,
-            lattice_spacing_a=lattice_spacing_a,
-            sigma_D=sigma_D,
-            alpha=90,
-            beta=90,
-            gamma=90,
-            symmetry=symmetry,
-            lattice_positions=lattice_positions,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types)
-
-    def symmetry_factor(self, h, k, l):
-        """Returns the symmetry factor (0 for forbidden)."""
-        return 1
-
-    def unit_cell_volume(self):
-        return self.lattice_spacing_a**3
-
-
-class BCCLattice(Lattice):
-    def __init__(self, objects, lattice_spacing_a=1.0, sigma_D=0.01):
-        ''' cannot specify lattice_types or lattice_coordinates'''
-        # Define the lattice
-        symmetry = {
-            'crystal family': 'cubic',
-            'crystal system': 'cubic',
-            'Bravais lattice': 'I',
-            'crystal class': 'hexoctahedral',
-            'point group': 'm3m',
-            'space group': 'Im3m',
-        }
-
-        lattice_coordinates = np.array([(0.0, 0.0, 0.0),
-                                        (0.5, 0.5, 0.5),
-                                        ])
-        lattice_types = [1, 1]
-        # positions = ['all']
-        lattice_positions = ['corner', 'center']
-
-        super(
-            BCCLattice,
-            self).__init__(
-            objects,
-            lattice_spacing_a=lattice_spacing_a,
-            sigma_D=sigma_D,
-            symmetry=symmetry,
-            lattice_positions=lattice_positions,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types)
-
-    def symmetry_factor(self, h, k, l):
-        """Returns the symmetry factor (0 for forbidden)."""
-        return 1
-        if (h + k + l) % 2 == 0:
-            return 2
-        else:
-            return 0
-
-    def unit_cell_volume(self):
-        return self.lattice_spacing_a**3
-
-    def q_hkl(self, h, k, l):
-        """Determines the position in reciprocal space for the given
-        reflection."""
-
-        prefactor = (2 * np.pi / self.lattice_spacing_a)
-        qhkl_vector = np.array([prefactor * h,
-                                prefactor * k,
-                                prefactor * l])
-        qhkl = np.sqrt(
-            qhkl_vector[0]**2 +
-            qhkl_vector[1]**2 +
-            qhkl_vector[2]**2)
-
-        return (qhkl, qhkl_vector)
-
-    def q_hkl_length(self, h, k, l):
-        prefactor = (2 * np.pi / self.lattice_spacing_a)
-        qhkl = prefactor * np.sqrt(h**2 + k**2 + l**2)
-
-        return qhkl
-
-
-class FCCLattice(Lattice):
-    def __init__(self, objects, lattice_spacing_a=1.0, sigma_D=0.01):
-        # Define the lattice
-        symmetry = {}
-        symmetry['crystal family'] = 'cubic'
-        symmetry['crystal system'] = 'cubic'
-        symmetry['Bravais lattice'] = 'F'
-        symmetry['crystal class'] = 'hexoctahedral'
-        symmetry['point group'] = 'm3m'
-        symmetry['space group'] = 'Fm3m'
-
-        # positions = ['all']
-        lattice_positions = ['corner', 'faceXY', 'faceYZ', 'faceXZ']
-
-        lattice_coordinates = [(0.0, 0.0, 0.0),
-                               (0.5, 0.5, 0.0),
-                               (0.0, 0.5, 0.5),
-                               (0.5, 0.0, 0.5),
-                               ]
-        lattice_types = [1, 1, 1, 1]
-
-        super(
-            FCCLattice,
-            self).__init__(
-            objects,
-            lattice_spacing_a=lattice_spacing_a,
-            sigma_D=sigma_D,
-            symmetry=symmetry,
-            lattice_positions=lattice_positions,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types)
-
-    def symmetry_factor(self, h, k, l):
-        """Returns the symmetry factor (0 for forbidden)."""
-        return 1
-        # if (h%2==0) and (k%2==0) and (l%2==0):
-        # All even
-        # return 4
-        # elif (h%2==1) and (k%2==1) and (l%2==1):
-        # All odd
-        # return 4
-        # else:
-        # return 0
-
-    def q_hkl(self, h, k, l):
-        """Determines the position in reciprocal space for the given
-        reflection."""
-
-        prefactor = (2 * np.pi / self.lattice_spacing_a)
-        qhkl_vector = np.array([prefactor * h,
-                                prefactor * k,
-                                prefactor * l])
-        qhkl = np.sqrt(
-            qhkl_vector[0]**2 +
-            qhkl_vector[1]**2 +
-            qhkl_vector[2]**2)
-
-        return (qhkl, qhkl_vector)
-
-    def q_hkl_length(self, h, k, l):
-        prefactor = (2 * np.pi / self.lattice_spacing_a)
-        qhkl = prefactor * np.sqrt(h**2 + k**2 + l**2)
-
-        return qhkl
-
-    def unit_cell_volume(self):
-
-        return self.lattice_spacing_a**3
-
-
-class DiamondTwoParticleLattice(Lattice):
-    # a.k.a. zincblende
-    '''
-        These are implied:
-            lattice_spacing_b = lattice_spacing_a
-            lattice_spacing_c = lattice_spacing_a
-            alpha = radians(90)
-            beta = radians(90)
-            gamma = radians(90)
-
-    '''
-
-    def __init__(self, objects, lattice_spacing_a=1.0, sigma_D=0.01):
-        lattice_spacing_a = lattice_spacing_a
-
-        sigma_D = np.array(sigma_D)          # Lattice disorder
-
-        # Define the lattice
-        symmetry = {
-            'crystal family': 'cubic',
-            'crystal system': 'cubic',
-            'Bravais lattice': 'P',
-            'crystal class': 'hexoctahedral',
-            'point group': 'm3m',
-            'space group': 'Pm3m',
-        }
-
-        # positions = ['cornerFCC', 'tetrahedralFCC']
-        lattice_positions = [
-            'corner',
-            'faceXY',
-            'faceYZ',
-            'faceXZ',
-            'tetra1',
-            'tetra2',
-            'tetra3',
-            'tetra4']
-
-        lattice_coordinates = [(0.0, 0.0, 0.0),
-                               (0.5, 0.5, 0.0),
-                               (0.0, 0.5, 0.5),
-                               (0.5, 0.0, 0.5),
-                               (0.25, 0.25, 0.25),
-                               (0.25, 0.75, 0.75),
-                               (0.75, 0.25, 0.75),
-                               (0.75, 0.75, 0.25),
-                               ]
-
-        lattice_types = [1, 1, 1, 1, 2, 2, 2, 2]
-
-        super(
-            DiamondTwoParticleLattice,
-            self).__init__(
-            objects,
-            lattice_spacing_a=lattice_spacing_a,
-            sigma_D=sigma_D,
-            symmetry=symmetry,
-            lattice_positions=lattice_positions,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types)
-
-    def symmetry_factor(self, h, k, l):
-        """Returns the symmetry factor (0 for forbidden)."""
-        return 1
-
-    def unit_cell_volume(self):
-        return self.lattice_spacing_a**3
-
-
-class RandomizedSimpleCubicLattice(Lattice):
-    def __init__(self, objects, lattice_spacing_a=1.0, sigma_D=0.01,
-                 filling_probs=None, lattice_coordinates=None,
-                 lattice_types=None, n_repeat=3):
-        ''' This is an extended lattice where we randomize the filling.
-            objects : the NanoObjects (or composite) to place in lattice
-            lattice_spacing_a/b/c : the a/b/c lattice spacings
-        '''
-        if filling_probs is None:
-            self.filling_probs = np.ones(len(objects))
-
-        # prepare variables of lattice
-        symmetry = {
-            'crystal family': 'cubic',
-            'crystal system': 'cubic',
-            'Bravais lattice': 'P',
-            'crystal class': 'hexoctahedral',
-            'point group': 'm3m',
-            'space group': 'Pm3m',
-        }
-
-        lattice_positions = ['corner'] * len(objects)  # not used
-
-        if lattice_coordinates is None:
-            lattice_coordinates = [(0.0, 0.0, 0.0)] * len(objects)
-        sub_lattice_coordinates = lattice_coordinates
-
-        lattice_coordinates = list()
-        lattice_types = list()
-        lattice_positions = list()
-        # now prepare the elements
-        for i in range(n_repeat):
-            x = i / n_repeat
-            for j in range(n_repeat):
-                y = j / n_repeat
-                for k in range(n_repeat):
-                    z = k / n_repeat
-                    for l, coord in enumerate(sub_lattice_coordinates):
-                        xsub, ysub, zsub = x + \
-                            coord[0], y + coord[1], z + coord[2]
-                        if np.random.uniform() <= filling_probs[l]:
-                            lattice_coordinates.append((xsub, ysub, zsub))
-                            lattice_types.append(l)
-                            lattice_positions.append("N/A")
-
-        lattice_spacing_a = n_repeat * lattice_spacing_a
-
-        # now call parent function to initialize
-        super(
-            RandomizedSimpleCubicLattice,
-            self).__init__(
-            objects,
-            lattice_spacing_a=lattice_spacing_a,
-            sigma_D=sigma_D,
-            alpha=90,
-            beta=90,
-            gamma=90,
-            symmetry=symmetry,
-            lattice_positions=lattice_positions,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types)
-
-    def symmetry_factor(self, h, k, l):
-        """Returns the symmetry factor (0 for forbidden)."""
-        return 1
-
-    def unit_cell_volume(self):
-        return self.lattice_spacing_a**3
-
-
-class ExtendedLattice(Lattice):
-    ''' An extended lattice of objects, can choose random fillings.
-        fill_probs : list of probabilities
-    '''
-
-    def __init__(
-            self,
-            objects,
-            lattice_spacing_a=1.0,
-            fill_probs=None,
-            lattice_spacing_b=None,
-            lattice_spacing_c=None,
-            alpha=90,
-            beta=None,
-            gamma=None,
-            sigma_D=0.01,
-            lattice_positions=None,
-            lattice_coordinates=None,
-            symmetry=None,
-            lattice_types=None,
-            repeat=3):
-        ''' For now only supports one type.'''
-        if fill_probs is None:
-            fill_probs = [1]
-
-        # sub_lattice_positions = lattice_positions
-        sub_lattice_coordinates = lattice_coordinates
-        # sub_lattice_types = lattice_types
-        if sub_lattice_coordinates is None:
-            sub_lattice_coordinates = [[0, 0, 0]]
-        lattice_positions = []
-        lattice_coordinates = []
-        lattice_objects = []
-        lattice_types = []
-
-        # fill_prob = fill_probs[0]
-
-        subunit_x, subunit_y, subunit_z = 1., 1., 1.
-        # TODO also add positions and types
-        for i in range(len(sub_lattice_coordinates)):
-            xp, yp, zp = sub_lattice_coordinates[i]
-
-            for ix in range(repeat):
-                xi = ix + xp * subunit_x
-                for iy in range(repeat):
-                    yi = iy + yp * subunit_y
-                    for iz in range(repeat):
-                        zi = iz + zp * subunit_z
-
-                        if(np.random.uniform(0.0, 1.0) <= fill_probs[i]):
-                            lattice_positions.append('N/A')
-                            lattice_types.append(i)
-                            lattice_coordinates.append((xi, yi, zi))
-                            lattice_objects.append(objects[i])
-
-        super(
-            ExtendedLattice,
-            self).__init__(
-            objects,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types,
-            lattice_positions=lattice_positions,
-            lattice_spacing_a=lattice_spacing_a,
-            lattice_spacing_b=lattice_spacing_b,
-            lattice_spacing_c=lattice_spacing_c,
-            alpha=alpha,
-            beta=beta,
-            gamma=gamma,
-            sigma_D=sigma_D)
-
-    def symmetry_factor(self, h, k, l):
-        """Returns the symmetry factor (0 for forbidden)."""
-        return 1
-
-    def unit_cell_volume(self):
-        return self.lattice_spacing_a**3
-
-
-class HexagonalLattice(Lattice):
-    def __init__(self, objects, lattice_spacing_a=1.0, lattice_spacing_b=None,
-                 lattice_spacing_c=None, sigma_D=0.01, lattice_types=None,
-                 lattice_coordinates=None, lattice_positions=None,
-                 symmetry=None, alpha=90, beta=90, gamma=60):
-        ''' Hexagonal lattice
-
-            lattice_spacing_a : the lattice spacing in the x plane
-            lattice_spacing_b : the lattice spacing in the x-y plane
-                if None, set to a
-            lattice_spacing_c : the lattice spacing along the z axis
-
-            Calculations:
-                lattice_spacing_a -> a
-                lattice_spacing_c -> c
-                a1 = (1, 0, 0)*a
-                a2 = (1/2., sqrt(3)/2., 0)*b
-                a3 = (0, 0, 1)*c
-
-                Reciprocal vectors:
-                b1 = 2*np.pi/a*(1, -1/np.sqrt(3), 0)
-                b2 = 2*np.pi/b*(0, 2/np.sqrt(3), 0)
-                b3 = 2*np.pi/c*(0, 0, 1)
-        '''
-        # Define the lattice
-        if symmetry is None:
-            symmetry = {}
-            symmetry['crystal family'] = 'cubic'
-            symmetry['crystal system'] = 'cubic'
-            symmetry['Bravais lattice'] = 'F'
-            symmetry['crystal class'] = 'hexoctahedral'
-            symmetry['point group'] = 'm3m'
-            symmetry['space group'] = 'Fm3m'
-
-        if lattice_positions is None:
-            lattice_positions = ['all']
-
-        if lattice_coordinates is None:
-            lattice_coordinates = [(0.0, 0.0, 0.0),
-                               ]
-        if lattice_types is None:
-            lattice_types = [1]
-        if lattice_spacing_b is None:
-            lattice_spacing_b = lattice_spacing_a
-        if lattice_spacing_c is None:
-            lattice_spacing_c = lattice_spacing_a
-
-        # TODO : gamma should be 60 deg I think (but alpha, beta, gamma are not
-        # really used)
-        super(
-            HexagonalLattice,
-            self).__init__(
-            objects,
-            lattice_spacing_a=lattice_spacing_a,
-            lattice_spacing_b=lattice_spacing_b,
-            lattice_spacing_c=lattice_spacing_c,
-            alpha=alpha,
-            beta=beta,
-            gamma=gamma,
-            sigma_D=sigma_D,
-            symmetry=symmetry,
-            lattice_positions=lattice_positions,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types)
-
-        # create the basis vectors and save
-        # TODO : make more similar to other implementations (should this go in
-        # super()?)
-        a = lattice_spacing_a
-        b = lattice_spacing_b
-        c = lattice_spacing_c
-
-        # the basis vectors for hexagonal
-        # TODO : should put this in init?
-        self.b1 = 2*np.pi/a*np.array([1, 1/np.sqrt(3), 0])
-        self.b2 = 2*np.pi/b*np.array([0, 2/np.sqrt(3), 0])
-        self.b3 = 2*np.pi/c*np.array([0, 0, 1])
-
-    def symmetry_factor(self, h, k, l):
-        """Returns the symmetry factor (0 for forbidden)."""
-        return 1
-        # if (h%2==0) and (k%2==0) and (l%2==0):
-        # All even
-        # return 4
-        # elif (h%2==1) and (k%2==1) and (l%2==1):
-        # All odd
-        # return 4
-        # else:
-        # return 0
-
-    def q_hkl(self, h, k, l):
-        """Determines the position in reciprocal space for the given reflection."""
-
-        qhkl_vector = h*self.b1 + k*self.b2 + l*self.b3
-
-        # NOTE: Valid for ideal hexagonal only
-        # NOTE : This was Kevin's old code, but should it not have been
-        # "-h + 2*k" instead of "h + 2*k"?
-        #qhkl_vector = ( 2*np.pi*h/(self.lattice_spacing_a), \
-                        #2*np.pi*(h+2*k)/(np.sqrt(3)*self.lattice_spacing_b), \
-                        #2*np.pi*l/(self.lattice_spacing_c) )
-        # I moved the 2*np.pi/lattice_spacing etc to the init in the variables
-        # self.b1, b2, b3 etc
-        qhkl = np.sqrt( qhkl_vector[0]**2 + qhkl_vector[1]**2 + qhkl_vector[2]**2 )
-
-        return (qhkl, qhkl_vector)
-
-
-
-    def q_hkl_length(self, h, k, l):
-        raise NotImplementedError("Add if needed")
-
-    def unit_cell_volume(self):
-        raise NotImplementedError("Add if needed")
-
-
-
-# Original author : Kevin G. Yager (most of these lattices are Kevin G. Yager)
-# HexagonalDiamondLattice
-###################################################################
-class HexagonalDiamondLattice(HexagonalLattice):
-    # Initialization
-    ########################################
-    def __init__(self, objects, lattice_spacing_a=None, lattice_spacing_b=None,
-                 lattice_spacing_c=None, alpha=90, beta=90, gamma=60,
-                 sigma_D=0.01):
-        ''' This is for a hexagonal diamond lattice.
-
-            Some useful relations:
-                lattice_spacing_a = bond_length*np.sqrt(2/3.)
-                where bond_length is the bond length. You must specify the
-                positionsi in terms of lattice spacing, not bond length so use
-                this conversion if needed.
-        '''
-        # write the lattice spacings in terms of the bond length
-        # this is different from the previous conventions but stresses
-        # that hexagonal diamond really depends on bond length
-
-        if lattice_spacing_a is None:
-            raise ValueError("Error, must specify at least the"
-                             " a lattice spacing.")
-        if lattice_spacing_b is None:
-            lattice_spacing_b = lattice_spacing_a
-        if lattice_spacing_c is None:
-            lattice_spacing_c = lattice_spacing_a*4.0/np.sqrt(6.)
-
-        # Define the lattice
-        symmetry = {}
-        symmetry['crystal family'] = 'triclinic'
-        symmetry['crystal system'] = 'triclinic'
-        symmetry['Bravais lattice'] = '?'
-        symmetry['crystal class'] = '?'
-        symmetry['point group'] = '?'
-        symmetry['space group'] = '?'
-
-        positions = ['network1', 'network2']
-        lattice_positions = ['corner bottom layer', \
-                                    'strut lower', \
-                                    'strut higher', \
-                                    'corner midlayer', \
-                                ]
-        # note: the vector components in x and y are 1/3 and
-        # not 1/2. This is because these vectors are non orthogonal
-        lattice_coordinates = [ (0.0, 0.0, 0.0), \
-                                 (1.0/3.0, 1.0/3.0, 1.0/8.0), \
-                                 (1.0/3.0, 1.0/3.0, 4.0/8.0), \
-                                 (0.0, 0.0, 5.0/8.0), \
-                               ]
-        lattice_objects = [objects[0], \
-                          ]
-                                    #objects[0], \
-                                    #objects[0], \
-                                    #objects[0], \
-                                #]
-        lattice_types = [1,1,1,1]
-
-        super(
-            HexagonalDiamondLattice,
-            self).__init__(
-            objects,
-            lattice_spacing_a=lattice_spacing_a,
-            lattice_spacing_b=lattice_spacing_b,
-            lattice_spacing_c=lattice_spacing_c,
-            sigma_D=sigma_D,
-            symmetry=symmetry,
-            lattice_positions=lattice_positions,
-            lattice_coordinates=lattice_coordinates,
-            lattice_types=lattice_types,
-            alpha=alpha,
-            beta=beta,
-            gamma=gamma)
